added support for block device mounts in policy framework
As part of the create_container enforcement point, we validate the mount points which includes validation of source, destination, type and options.
While the destination, type and options can be validated as per the existing logic, we need to validate the source separately for block mounts.
This change adds the support for the same.

Signed-off-by: Harsh Rawat <harshrawat@microsoft.com>