add validation to prevent mounting VHDs at arbitrary location within UVM
Prior to this change, whenever we mounted any VHD onto the UVM, the path for
the mount within the UVM was not validated. This meant that if the host
gets compromised, an attacker can request the mount to be added at arbitrary
location within the host. This change adds the check which ensures that the requested mount path is
at an expected location.

Additionally, we have introduced the mechanism wherein a VHD for confidential container mount
can only be mounted as a block device.

Signed-off-by: Harsh Rawat <harshrawat@microsoft.com>