Merged PR 12588213: Various enforcement fixes
[Cherry-picked from ba849d8464ff019fe751872d458168bc9aec2256]

While the entire PR contains several (perhaps unrelated) changes, the individual commits in this PR can be reviewed on their own, and there is some background in the commit messages.

I've tested this with azcri-containerd as well as with the functional tests in hcsshim and some of the tests fail due to unrelated reasons (can't do VPMem, no GPU, WCOW) - I haven't noticed anything that could be caused by these changes. I also deployed these changes to a VM and can run some confidential workloads manually.

Outstanding work that will not be done in this PR:

- Deny unmounting of in-use layers (this should already be impossible due to directories being in use, but can result in rego metadata being inconsistent)

Missing go tests that will not be part of this PR - I have local stash for these, but they are not fully working. Can do separately:

- Functional test for checking that a bad OCISpec.Rootfs is rejected (tested manually)
- Functional test for checking that a bad containerID is rejected (tested manually)

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>